Add files via upload
Day 13 was creating a Higher-Lower game.

In some ways, my code is more elegant than the instructional example. In other ways...not so much. I feel like there's a more efficient way to compare the player choice against the competition than the way I did it...but it works, for now.