Add files via upload
Day 16, and our introduction to OOP. I got 95% of this before looking for help, and I'm glad I did: I misread the "money_machine" code and it was causing me a headache. Also, I needed to pass around a variable, which in retrospect was a bit of a "durr" moment.