Add files via upload
Day 22, a game of Pong.

I was pretty pleased with the fact that I was able to setup 95% of this on my own, an even put in a neat little center line. I had originally placed a great deal of the Screen setup in a separate file to import, but later slapped it in Main. I learned a lot about what to import into where (OOP stuff), so that was good (for instance, I had duplicated the Scoreboard and was going crazy trying to figure out why the .clear() wasn't working. Derp.).

My original method for moving the ball was messy and complicated and really didn't work. Also, getting the paddles to recognize the ball was something that I was never going to figure out without help. So it goes.