Add files via upload
Day 24 was all about reading and writing from files. The project for the day was to create a simple mail merge.

This project went swimmingly, though a not-insubstantial portion of the instructor's instruction was "Google it," which I don't love.

Thankfully my Google Fu is strong, and I cranked this out in no-time-flat.