Add files via upload
Day 28, making a timer.

This exercise was a little strange, as I feel like there must be a better way to make a timer. Also, a lot of handholding. Also also, she started using a lot of global variables which we never really went over too much since we were told...y'know...NOT to use them. I feel like this was a program that would have benefited from our previously focused-on OOP method.

Alright, needless bitching aside, I found the GUI creation to be a breeze. Feel very solid on that front.