Add files via upload
Day 32.
The hardest part of this lesson was, once again, dictionaries and DataFrames. My problem was reading the "hint" the instructor gave for how to import and organize the CSV file, which was unnecessarily convoluted. Once I ignored that and figured out how to organize the dictionary properly, everything fell into place.

ALL of the code below is my own and it works great, which feels nice after yesterday's catastrophe.

I'm pretty sure I could organize the Letters better, but eh.