Add files via upload
Day 33.

Today we messed about with API pulls, which is kind of exciting.

The app is supposed to e-mail me whenever the International Space Station is overhead. As far as I can tell, it works! My version is touch different than the instructors, as I ran everything through functions which then looped into the "main" function (i.e., ship_it()). 

Some of this code was done throughout the lesson today, so not all of it is mine (all the Functions are, though). I feel like there is probably an easier way to extract the Hour data from the ISS API than to split the data a bunch of times, but I also didn't make the effort to find out what that easier way is. So.

I seem to be getting a handle on dictionaries, and I'm kinda thrilled to be working with APIs.