Add files via upload
Day 34.

We updated the old Quizzler program we made in days past, this time using API and a GUI. It now has buttons and things. Tre fancy.

So, the original Quizzler program was dizzying in its complexity, what with passing things in and out of other things using OOP standards, and this exercise didn't reduce that in any way. There was a lot of hand-holding this lesson.

Also, I learned that TkInter has some really obscure error coding?